Rename pyIMS to pyImagingMSpec
