+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. pyIMS documentation master file, created by
+.. pyImagingMSpec documentation master file, created by
    sphinx-quickstart on Tue Feb  9 15:19:18 2016.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to pyIMS's documentation!
+Welcome to pyImagingMSpec's documentation!
 =================================
 
 .. toctree::
