+++ pyImagingMSpec/convert/h5.py
