+++ pyImagingMSpec/convert/hdf5.py
