+++ pyImagingMSpec/convert/sl.py
