+++ pyImagingMSpec/smoothing.py
