+++ setup.py
@@ -1,10 +1,10 @@
 from setuptools import setup, find_packages
 
-from pyIMS import __version__
+from pyImagingMSpec import __version__
 
-setup(name='pyIMS',
+setup(name='pyImagingMSpec',
       version=__version__,
       description='Python library for processing imaging mass spectrometry data',
-      url='https://github.com/alexandrovteam/pyIMS',
+      url='https://github.com/alexandrovteam/pyImagingMSpec',
       author='Alexandrov Team, EMBL',
       packages=find_packages())
